+++ ignite/templates/module/placeholders.go
@@ -5,7 +5,7 @@ const (
 	Placeholder2 = "// this line is used by starport scaffolding # 2"
 	Placeholder3 = "// this line is used by starport scaffolding # 3"
 
-	// Placeholders in Stargate app.go
+	// Placeholders in app.go
 	PlaceholderSgAppModuleImport        = "// this line is used by starport scaffolding # stargate/app/moduleImport"
 	PlaceholderSgAppModuleBasic         = "// this line is used by starport scaffolding # stargate/app/moduleBasic"
 	PlaceholderSgAppKeeperDeclaration   = "// this line is used by starport scaffolding # stargate/app/keeperDeclaration"
@@ -21,7 +21,7 @@ const (
 	PlaceholderSgAppBeforeInitReturn    = "// this line is used by starport scaffolding # stargate/app/beforeInitReturn"
 	PlaceholderSgAppMaccPerms           = "// this line is used by starport scaffolding # stargate/app/maccPerms"
 
-	// Placeholders in Stargate app.go for wasm
+	// Placeholders in app.go for wasm
 	PlaceholderSgWasmAppEnabledProposals = "// this line is used by starport scaffolding # stargate/wasm/app/enabledProposals"
 
 	// Placeholders in cmd/appd/cmd/root.go
